Adding "relative mode" to OnScreenStick
Currently, the controller uses your first touch to determine a "start point". In many cases its actually helpful to always have the "start point" at the center. 
Scenario, if I tap right on the controller, currently this would not register any movement because the origin of the movement is reset to this point. Ideally if I tap right, I would expect the controller to register this movement.