Handle parent directory renaming whilst asset is being edited. (#1049)
* Added test to ensure that an exception is not thrown when Saving input asset and the parent directory was renamed. (case 1207527)

* Removed cached asset path. We now use the asset guid to get the path when requested. (case 1207527)

When the parent directory is renamed the path becomes invalid. Its much safer to use the asset guid and resolve the path when its requested. There is no continuous use of the path in code so it should not have any performance impact.

* Code Formatting

* Discard the unused argument.

Fixes a warning.

* Ignore warning CA1801

* Test the file has been modified instead of an exception not being thrown.

We should test for the behaviour we actually want.

* Added CHANGELOG entry

* Use the [Unreleased] header instead of a unreleased version and date