Fixed Hold Interaction not working with multiple bindings. (case 1239551)
The problem was in how we copy over the state. If the action already had a magnitude then we would copy the flags but not the magnitude which caused the value to be wrong. We need to calculate the magnitude and then copy over the flags so that the value is always correct.