FIX: PlayerInputManager join action not triggering
This fixes the following case. An input action asset defines an action called Fire that uses the space bar key and the gamepad south button. The PlayerInputManager join action is a reference to the Fire action. The first player hits the space bar key. This associates the keyboard with that user, and *disables input events from the gamepad* for the current action map. The second player now tries to join by pressing the gamepad south button, but that input is no longer active, so they can't join. This only affects PlayerInputManager when:
  * join behaviour is set to Join Players When Join Action Is Triggered
  * the join action is a reference