Update InputSystemUIInputModule.cs
Imagine a panel containing a single selected button, where pressing the escape key should dismiss the panel. The most intuitive way to implement this is to wire up Esc with Cancel and let the EventSystem do the work. Button, which is the currently selected object, doesn't implement ICancelHandler, so naturally it should let another UI element higher up in the hierarchy, a script attached to the panel, do the work.

This currently doesn't happen, as the event is not propagated up the hierarchy. This PR adds an option (default false for compatibility) to surface IMoveHandler, ISubmitHandler and ICancelHandler up the hierarchy for a custom UI element to handle it.