Replace AutoDisposeObserver with something simpler
The idea of disposing the subscriber all the way up the subscriber tree that CallOnce uses is not great here and will lead to bugs and confusion. If the user creates an observable from InputEventListener, and then attaches multiple subscribers where one is a CallOnce call, all the other subscribers will also stop getting events when the CallOnce disposes the source, which is not great. I don't have a good solution for this yet though.