FIX: Input from other devices not working when DualShock 4 controller is connected
The sensors on the DualShock 4 have not yet been mapped to controls and marked as noisy, so whenever those sensors send events, which is all the time, the system has no way to know that the events represent noisy data. This means the DualShock 4 constantly sets itself as the current device, even if the user is using a different device.