Improve dirty asset detection when new asset is first created
Ideally at this stage, the file on disk should have a serializable equivalent as the serialized asset. However, because the new asset on disk is being created with a specific default json layout, the serialized JSON asset doesn't have not the same content.
There are opportunities for change in this area.