NEW: Allow selection of InputActionReferences from project-wide actions (#1779)
* Add InputActionReference advanced picker

This new picker allows us to search InputActionReferences from the project wide actions asset. Since this is asset is in ProjectSettings, a normal object picker doesn't work.

* Update InputActionReferences of project-wide actions on save

* Check for dangling action references in Inspector and set them to "None"

A dangling action reference means a actin reference that point to a non-existent InputAction.
This needs to be checked here in case an input action is removed from the project-wide actions asset. Otherwise the UI will still hold on to a dangling action reference.

* Remove unused code

* Fix comments references

* Correct variables naming

* Test updating references when asset actions are changed

* Only check dangling project-wide action references

* Update CHANGELOG

* Make InputActionAssetIconLoader methods internal

* Make advanced picker open by default with list view and simple search query

OpenInBuilder mode allows to search by "tags". But no implementation was done so they did not work.

* Run added tests only on the Editor