Make `DisableAndResetAllDevices` handle events
Input events queued when the player runs in the background weren't being handled before, when setting the background behavior with `DisableAndResetAllDevices` option. This would result in A LOT of events needed to be processed once the player regained focus. Also,action callbacks could be invoked incorrectly depending on what events were queued in the background.