CHANGE: Override UGUI EventSystem with InputSystemUIInputModule (#1902)
* Override InputModule component of EventSystem

* Make it override only when ISX is enabled and IM disabled

InputSystem should override the input module for the `EventSystem` game object only when it's the only input system available. If IM (Input Manager) is available as well or is the only input backend available, we fallback to default behavior of adding `StandaloneInputModule`.

* Use static ObjectFactory.AddComponent

* Test that UGUI EventSystem is created correctly

This test assumes that the Input System package is enabled in the
project. In cases like this, the EventSystem GameObject will always
be created with the InputSystemUIInputModule.

* Rely on ENABLE_INPUT_SYSTEM instead

This avoids us having to use `EditorPlayerSettingsHelpers` which uses `FindObjectsOfTypeAll`. This is not allowed to be done on ScriptableObject constructor (or instance field initializer)

* Update CHANGELOG

---------

Co-authored-by: Lyndon Homewood <33493311+lyndon-unity@users.noreply.github.com>