Set left and right sticks default state to avoid test failures
The default state for zero value of sticks was not being set through default state.

We only set a defaultState if there's at least a control with defaultState set. Otherwise, we seem to be bypassing it. For some reason, it seems we were deriving the default state of the sticks through the parameters control have but I cannot say for sure without spending much more time on this.